fix: 完成账号显示模板设置页面

+++ app/src/main/java/vip/mystery0/xhu/timetable/viewmodel/MainViewModel.kt
@@ -128,8 +128,8 @@ class MainViewModel : ComposeViewModel() {
     private val _backgroundImage = MutableStateFlow<Any>(Unit)
     val backgroundImage: StateFlow<Any> = _backgroundImage
 
-    private val _backgroundImageBlur = MutableStateFlow(1)
-    val backgroundImageBlur: StateFlow<Int> = _backgroundImageBlur
+    private val _backgroundImageBlur = MutableStateFlow(1F)
+    val backgroundImageBlur: StateFlow<Float> = _backgroundImageBlur
 
     private val _emptyUser = MutableStateFlow(false)
     val emptyUser: StateFlow<Boolean> = _emptyUser
